Phase C (B2): signed Characteristic-points budget (Weak Characteristics)
Effect::CharacteristicPoints.amount u8 -> i8 and characteristic_points_granted
now returns i32, so a grant may be negative. Weak Characteristics removes 3
budget points (Core:7056-7058, takeable twice); it nets against Improved
Characteristics (+3). DerivedScores field widened to i32; the UI already sums it
into start_points, so negatives just lower the budget.

TDD: effective.rs test for the negative grant + netting. Workspace tests,
clippy, fmt green.

+++ crates/arm-rules/src/validation.rs
@@ -1730,7 +1730,7 @@ fn validate_characteristics(entity: &Entity, ruleset: &Ruleset, issues: &mut Vec
     // Improved Characteristics (+3 each, stackable) raises the buy budget above
     // the ruleset's base start_points.
     let granted = crate::effective::characteristic_points_granted(entity, ruleset);
-    let budget = i32::from(rules.start_points) + granted as i32;
+    let budget = i32::from(rules.start_points) + granted;
     if cost > budget {
         issues.push(ValidationIssue::error(
             ValidationIssue::CODE_CHARACTERISTIC_OVERSPENT,
